Move PersistentStorage in separate file
