+++ common/credential-storage/src/backends/mod.rs
@@ -1,0 +1,2 @@
+// Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
